requirements.txt
# Excel Translator App Dependencies

# Core GUI Framework
PyQt6>=6.5.0

# Excel file processing
openpyxl>=3.1.0
xlrd>=2.0.1

# DeepL API client
deepl>=1.13.0

# HTTP requests and utilities
requests>=2.31.0
urllib3>=2.0.4

# Data processing
pandas>=2.0.0

# Logging and configuration
python-json-logger>=2.0.7

# Additional utilities
Pillow>=10.0.0
chardet>=5.1.0

# Development dependencies (optional)
pytest>=7.4.0
pytest-qt>=4.2.0
pyinstaller>=5.13.0

# System dependencies
psutil>=5.9.0
